requirements.txt
discord.py==1.3.3
dateparser==0.7.4